% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_clean_data.R
\name{add_all_hours}
\alias{add_all_hours}
\title{Add all hours between min and max date}
\usage{
add_all_hours(dat)
}
\arguments{
\item{dat}{data.frame with columns station_id (must be unique) and timestamp}
}
\description{
Sometimes rows are completely missing when no single value is provided.
If rows are removed after our first import, the next import would not remove those rows
because it could only overwrite existing rows. So we need a row with NAs for all missing hours.
Since we always import complete days, we insert hours for all days (from 00:00 of the first to
23:00 of the last day)
}
