% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_download_links.R
\name{add_download_links_for_year}
\alias{add_download_links_for_year}
\title{Get urls to hourly data website and their csv file for the specified year}
\usage{
add_download_links_for_year(station_info_df, year = "2015")
}
\arguments{
\item{station_info_df}{data.frame;}

\item{year}{character;}
}
\value{
data.frame;
}
\description{
for each station in the station info data frame and append them.
Builds link to csv differently because of
\url{https://github.com/INWT/fairq-data-messstationen/issues/9}.
}
