% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_download_links.R
\name{add_download_links_last_2_months}
\alias{add_download_links_last_2_months}
\title{Get urls to hourly data csv file for the last two months
for each station in the station info data frame and append them.}
\usage{
add_download_links_last_2_months(station_info_df)
}
\arguments{
\item{station_info_df}{data.frame;}
}
\value{
data.frame;
}
\description{
Get urls to hourly data csv file for the last two months
for each station in the station info data frame and append them.
}
