% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_station_info.R
\name{add_missing_karl_marx_entry}
\alias{add_missing_karl_marx_entry}
\title{Add missing entry for Karl-Marx-Str.}
\usage{
add_missing_karl_marx_entry(station_info)
}
\arguments{
\item{station_info}{data frame with columns station_name, station_id and station_url, containing
all stations crawled from the website}
}
\value{
data frame with additional row. If the original data frame contained ID 220, we add ID
221 and vice versa.
}
\description{
The Karl-Marx-Str. station (ID 220) is being replaced temporarily due to construction works.
The replacement station has the ID 221.
No matter which date is crawled, we want both IDs to be in the list of crawled stations.
Source: https://luftdaten.berlin.de/station/mc221#station-info
}
