% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_clean_data.R
\name{clean_station_csv}
\alias{clean_station_csv}
\title{Clean station csv file}
\usage{
clean_station_csv(station_csv, id, hourly = TRUE)
}
\arguments{
\item{station_csv}{data.frame;}

\item{id}{character; station id}

\item{hourly}{should hourly or daily data be cleaned, default: hourly}
}
\value{
data.frame
}
\description{
Make header, clean names and add station id as column
}
