% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_station_info.R
\name{create_station_info}
\alias{create_station_info}
\title{Wrapper to create mapping table with station id, name & url.}
\usage{
create_station_info(url = "https://luftdaten.berlin.de")
}
\arguments{
\item{url}{character;}
}
\value{
data.frame
}
\description{
Wrapper to create mapping table with station id, name & url.
}
