% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_download_csvs.R
\name{get_station_csv}
\alias{get_station_csv}
\title{Wrapper to read station csv files, clean them and add station id as column.}
\usage{
get_station_csv(station_info_df, hourly = TRUE)
}
\arguments{
\item{station_info_df}{data.frame;}

\item{hourly}{TRUE per default files with hourly data are extracted and cleaned}
}
\value{
list;
}
\description{
Wrapper to read station csv files, clean them and add station id as column.
}
