% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_station_info.R
\name{get_station_id}
\alias{get_station_id}
\title{Get each station's id}
\usage{
get_station_id(table_row)
}
\arguments{
\item{table_row}{html_node; html table row with station info}
}
\value{
character;
}
\description{
Get the text element of the table row (<tr>) and get the station id
}
