% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_station_info.R
\name{get_station_name}
\alias{get_station_name}
\title{Get each station's name}
\usage{
get_station_name(table_row)
}
\arguments{
\item{table_row}{html_node; html table row with station info}
}
\value{
character;
}
\description{
Get the text element of the table row (<tr>) and get the station name.
}
