% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_station_info.R
\name{get_station_table}
\alias{get_station_table}
\title{Crawls station table and returns html elements.}
\usage{
get_station_table(url)
}
\arguments{
\item{url}{character;}
}
\value{
xml_nodeset;
}
\description{
Requests the body of the station table and returns the html
elements within, which are all the table rows (<tr>).
}
