% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_station_info.R
\name{get_station_url}
\alias{get_station_url}
\title{Get each station's url}
\usage{
get_station_url(table_row, url)
}
\arguments{
\item{table_row}{html_node; html table row with station info}

\item{url}{character;}
}
\value{
character;
}
\description{
Go into button element of the table row (<tr>) and get the relative path to station url,
then paste it with the base url.
}
