% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_clean_data.R
\name{manage_columns}
\alias{manage_columns}
\title{Preserve the table columns and their order.}
\usage{
manage_columns(df)
}
\arguments{
\item{df}{data.frame;}
}
\value{
data.frame;
}
\description{
All db table columns that are not present in the df are added
and then ordered according to the db table.
}
