% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_download_csvs.R
\name{read_station_csv}
\alias{read_station_csv}
\title{Read station csv file}
\usage{
read_station_csv(station_info_df, id)
}
\arguments{
\item{station_info_df}{data.frame;}

\item{id}{character; station id}
}
\value{
data.frame
}
\description{
Read station csv file
}
